lab5/requirements.txt
cryptography==3.2.1
Django==3.1.3
django-cors-headers==3.5.0
django-crispy-forms==1.10.0
django-storages==1.10.1
djangorestframework==3.12.2
django-extensions==3.0.9
psycopg2==2.8.6
pysaml2==4.5.0
argon2-cffi==20.1.0